GraphQLUnion initial (somewhat incomplete) support
Equality checking for union types (which has to rely on checking for disjoint
collections [Collections.disjoin], which, in turn relies on #equals) doesn't
work in cases when multiple instance of the same type are produced (which is
a current issue in graphql-java-annotations), so a couple of wrappers were added
to rely on some basic equality tests (like names). It's not a great solution,
but works as a temp workaround.