Added support for method Batched Data Fetcher
This change allows methods to be used as Batched data fetchers.
Now it's possible to write batched queries like this:

    public class SimpleBatchedField {
        @GraphQLField
        @GraphQLBatched
        public static List<String> a() {
            return Arrays.asList("one", "two");
        }
    }

A method witht the @GraphQLBatched annotation should alwasy return
a list. It also has to be static, because it won't operate on a single
object but a list of objects, which should be obtained from the
DataFetchingEnvironment.