private field annotation was still broken, fixed it and added data fetching tests for it also. (#28)
I thought that it was using propertyDataGetter for fields (what fallbacks to field), but it was using fieldDataGetter for fields (which is faster).
Added some logic which is trying to decide which one to use.

Basic idea on this change is that, if there is getter, use it. Otherwise use field directly.