Make it possible to provide custom prefixes AND postfixes for input object name generation. This is setup on the container. The defaults were not changed and are equivalent to : container.setInputPrefix("Input"); container.setInputPostfix("";
If you want to change the naming generation to something more "compliant" with GraphQL's naming conventions you can change the configuration to:

container.setInputPrefix("");
container.setInputPostfix("Input");

So for example if generating a name for the input type generated from the class Books it will be called InputBooks in the first configuration and BooksInput in the second configuration. This is also very useful if you are using prefixes to namespace your classes such as : ACMEBooks will be InputACMEBooks in the first configuration and ACMEBooksInput in the second configuration.

As you can see the prefix and postfix configuration is also very flexible and the "Input" word is just a configuration variable now.

By default however, nothing changes so existing projects should not break.

Some unit tests were also added to make sure that the generation works properly.